requirements.txt
# Core LangChain ecosystem
langchain
langchain-community
langchain-core

# Embeddings & Vector DB
faiss-cpu
sentence-transformers

# Text processing & loaders
unstructured
pypdf
python-docx
docx2txt
pandas
openpyxl
rank-bm25

# Web loading
requests
beautifulsoup4
# Config & Utilities
PyYAML
tqdm

langchain-groq
python-dotenv

Flask
flask-cors